# web/eval_factory_console/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentShellSessionPageV1": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "schema_version": {
            "const": "eval-factory/agent-shell-session-page/v1",
            "default": "eval-factory/agent-shell-session-page/v1",
            "title": "Schema Version",
            "type": "string"
          },
          "sessions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AgentShellSessionSummaryV1"
            },
            "maxItems": 500,
            "title": "Sessions",
            "type": "array"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "offset",
          "limit",
          "total"
        ],
        "title": "AgentShellSessionPageV1",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Conversation-first Harness session API with committed sequence SSE, bounded source admission, and owner-backed service composition.",
    "title": "Eval Dataset Factory Agent API",
    "version": "1.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/harness/sessions": {
      "get": {
        "operationId": "list_sessions_api_harness_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentShellSessionPageV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions"
      },
      "post": {
        "operationId": "create_session_api_harness_sessions_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Eval-Factory-Principal",
            "required": true,
            "schema": {
              "title": "X-Eval-Factory-Principal",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "idempotency_key": {
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9._:/-]{2,255}$",
                    "title": "Idempotency Key",
                    "type": "string"
                  },
                  "incarnation_id": {
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9._:/-]{2,255}$",
                    "title": "Incarnation Id",
                    "type": "string"
                  },
                  "schema_version": {
                    "const": "eval-factory/agent-shell-create-session-command/v1",
                    "default": "eval-factory/agent-shell-create-session-command/v1",
                    "title": "Schema Version",
                    "type": "string"
                  },
                  "session_id": {
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9._:/-]{2,255}$",
                    "title": "Session Id",
                    "type": "string"
                  }
                },
                "required": [
                  "session_id",
                  "incarnation_id",
                  "idempotency_key"
                ],
                "title": "AgentShellCreateSessionCommandV1",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentShellProjectionV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session"
      }
    },
    "/api/harness/sessions/{session_id}": {
      "delete": {
        "operationId": "close_session_api_harness_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Eval-Factory-Principal",
            "required": true,
            "schema": {
              "title": "X-Eval-Factory-Principal",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "expected_session_version": {
                    "maximum": 1000000000,
                    "minimum": 1,
                    "title": "Expected Session Version",
                    "type": "integer"
                  },
                  "idempotency_key": {
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9._:/-]{2,255}$",
                    "title": "Idempotency Key",
                    "type": "string"
                  },
                  "schema_version": {
                    "const": "eval-factory/agent-shell-close-session-command/v1",
                    "default": "eval-factory/agent-shell-close-session-command/v1",
                    "title": "Schema Version",
                    "type": "string"
                  }
                },
                "required": [
                  "expected_session_version",
                  "idempotency_key"
                ],
                "title": "AgentShellCloseSessionCommandV1",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentShellProjectionV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Session"
      },
      "get": {
        "operationId": "get_session_api_harness_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentShellProjectionV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session"
      }
    },
...
    "/api/harness/sessions/{session_id}/events": {
      "get": {
        "operationId": "list_events_api_harness_sessions__session_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_sequence",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Sequence",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentShellEventPageV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events"
      }
    },
    "/api/harness/sessions/{session_id}/members/{member_id}": {
      "get": {
        "operationId": "get_member_api_harness_sessions__session_id__members__member_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "member_id",
            "required": true,
            "schema": {
              "title": "Member Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentShellMemberProjectionV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Member"
      }
    },
...
    "/api/harness/sessions/{session_id}/team": {
      "get": {
        "operationId": "get_team_api_harness_sessions__session_id__team_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentShellTeamSummaryV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Team"
      }
    },
...
  }
}